+++ ZenLibBench/AclBench.cs
@@ -69,7 +69,7 @@ namespace ZenLibBench
         public void VerifyAclProvenance()
         {
             var f = Function<IpHeader, (bool, ushort)>(h => this.acl.ProcessProvenance(h));
-            var packet = f.Find((p, o) => o.Item2() == (this.acl.Lines.Length + 1), backend: this.Backend);
+            var packet = f.Find((p, o) => o.Item2() == (ushort)(this.acl.Lines.Length + 1), backend: this.Backend);
             f.Evaluate(packet.Value);
         }
     }
