+++ ZenLibTests/Network/Acl.cs
@@ -45,12 +45,12 @@ namespace ZenLib.Tests.Network
         /// <returns>Whether permitted.</returns>
         public Zen<(bool, ushort)> ProcessProvenance(Zen<IpHeader> hdr)
         {
-            var acc = ValueTuple<bool, ushort>(false, this.Lines.Length + 1);
+            var acc = ValueTuple<bool, ushort>(false, (ushort)(this.Lines.Length + 1));
 
             for (int i = this.Lines.Length - 1; i >= 0; i--)
             {
                 var line = this.Lines[i];
-                acc = If(line.Matches(hdr), ValueTuple<bool, ushort>(line.Permitted, i + 1), acc);
+                acc = If(line.Matches(hdr), ValueTuple<bool, ushort>(line.Permitted, (ushort)(i + 1)), acc);
             }
 
             return acc;
@@ -66,13 +66,13 @@ namespace ZenLib.Tests.Network
         {
             if (i == this.Lines.Length)
             {
-                return ValueTuple<bool, ushort>(false, i + 1);
+                return ValueTuple<bool, ushort>(false, (ushort)(i + 1));
             }
 
             var line = this.Lines[i];
             return If(
                 line.Matches(hdr),
-                ValueTuple<bool, ushort>(line.Permitted, i + 1),
+                ValueTuple<bool, ushort>(line.Permitted, (ushort)(i + 1)),
                 ProcessProvenance(hdr, i + 1));
         }
